feat(api): faster_decoding / container / progressive_dc / premultiply (W4-3)
Adds four libjxl-parity knobs as builders on LossyConfig / LosslessConfig
(plus a matching one on EncodeRequest) and the corresponding cjxl-rs CLI
flags:

- with_faster_decoding(u8) / --faster-decoding 0..4
    Mirrors libjxl cparams.decoding_speed_tier
    (enc_frame.cc:280, enc_modular.cc:469-516,
    enc_ac_strategy.cc:884, enc_ans.cc:1372). Per-tier semantics are
    documented on the builder rustdoc. Values clamp to
    MAX_FASTER_DECODING = 4.

- with_container_mode(ContainerMode) / --container -1|0|1
    Mirrors libjxl cjxl --container 0|1. New ContainerMode enum:
    Auto (default, wrap on metadata or codestream-level demand),
    Always (force wrap), Never (bare codestream, drop metadata).

- with_progressive_dc(u8) / --progressive-dc 0..2 (lossy only)
    Mirrors libjxl cjxl --progressive_dc 0..2. Level 1 implies
    with_lf_frame(true) and produces byte-identical output to the
    existing --lf-frame flag (verified on a CLIC photo:
    854 vs 854 bytes, diff returns 0). Level 2 is stored for forward
    compatibility (the encoder currently emits a single LfFrame).
    Values clamp to MAX_PROGRESSIVE_DC = 2.

- with_premultiplied_alpha_mode(PremultipliedAlphaMode) / --premultiply -1|0|1
    Mirrors libjxl cjxl --premultiply -1|0|1. New
    PremultipliedAlphaMode enum: Off / On / Auto. The boolean
    with_premultiplied_alpha() is the existing path and still works;
    setting the mode overrides it. Auto resolution at encode time is
    queued follow-on work.

Cleanups (pre-existing clippy items flagged in the W3-3 audit):

- effort.rs:1717 — drop unnecessary same-type cast
  (`p11.butteraugli_iters as u32 <= crate::api::MAX_QUANT_LOOP_ITERS`).
- modular/inline_add_sample.rs:457 — drop unnecessary same-type
  cast (`b.add_prop_i32(i as i32).unwrap()`).

Tests: 5 new unit tests in api::tests cover defaults, clamping,
builder round-trip, and the progressive_dc>=1 => lf_frame
implication. cargo test -p jxl-encoder --lib: 1083 passed,
0 failed.

CLI smoke verifies all four flags parse + run on lossy + lossless +
RGBA inputs. The faster-decoding and container knobs are stored on
the config and surfaced via getters — full encoder-side consumption
is queued follow-on work (matching the store-and-surface pattern
used by the previous A1 audit additions in 7ce49bc0).

+++ jxl-encoder/src/effort.rs
@@ -1913,7 +1913,7 @@ mod tests {
         // (The lossy() clamp pins at 11; verify the table never returns
         // anything above the loop's structural cap.)
         assert!(
-            p11.butteraugli_iters as u32 <= crate::api::MAX_QUANT_LOOP_ITERS,
+            p11.butteraugli_iters <= crate::api::MAX_QUANT_LOOP_ITERS,
             "butteraugli_iters must not exceed MAX_QUANT_LOOP_ITERS"
         );
     }
